+++ tests/integration/cli/test_export_inference.py
@@ -48,6 +48,8 @@ TASK_NAME_TO_MAIN_METRIC_NAME = {
     "visual_prompting": "test/f1-score",
     "zero_shot_visual_prompting": "test/f1-score",
     "action_classification": "test/accuracy",
+    "keypoint_detection": "test/PCK",
+    "object_detection_3d": "test/AP_3d@0.5",
 }